requirements.txt
streamlit
langchain
python-dotenv
serpapi
openai
